+++ tests/LaravelModulesServiceProviderTest.php
@@ -4,6 +4,7 @@ namespace Nwidart\Modules\Tests;
 
 use Nwidart\Modules\Contracts\ActivatorInterface;
 use Nwidart\Modules\Contracts\RepositoryInterface;
+use Nwidart\Modules\Exceptions\InvalidActivatorClass;
 
 class LaravelModulesServiceProviderTest extends BaseTestCase
 {
@@ -17,6 +18,16 @@ class LaravelModulesServiceProviderTest extends BaseTestCase
     /** @test */
     public function it_binds_activator_to_activator_class()
     {
+        $this->assertInstanceOf(ActivatorInterface::class, app(ActivatorInterface::class));
+    }
+
+    /** @test */
+    public function it_throws_exception_if_config_is_invalid()
+    {
+        $this->expectException(InvalidActivatorClass::class);
+
+        $this->app['config']->set('modules.activators.file', ['class' => null]);
+
         $this->assertInstanceOf(ActivatorInterface::class, app(ActivatorInterface::class));
     }
 }